pyproject.toml
[project]
name = "secxxii-telegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.9",
    "langchain-groq>=0.2.1",
    "langchain-openai>=0.2.10",
    "langgraph>=0.2.53",
    "psycopg>=3.2.3",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.36",
]
